Module tydi::design

Constructs that are used to generate hardware designs, that are not part of the specification (yet).

Re-exports

pub use library::Library;

pub use project::Project;

pub use streamlet::Interface;

pub use streamlet::Mode;

pub use streamlet::Streamlet;

Modules

implementation

library
This module contains the Library structure, used to group multiple Streamlets together.

param

project

streamlet
This module contains the Streamlet structure.

Structs

NodeIFHandle

ParamHandle

StreamletHandle
Handles for objects inside a project, through project hierarchy

Constants

GEN_LIB

THIS_KEY

Type Definitions

ComponentKey

IFKey

LibKey
Index types

NodeKey

ParamKey

ParamStoreKey

StreamletKey