Trait tydi::generator::common::convert::ModeFor

pub trait ModeFor {
    fn mode_for(&self, streamlet_mode: Mode) -> Mode;
}

Trait that helps to determine the common representation port mode given a streamlet interface mode.

Required methods

fn mode_for(&self, streamlet_mode: Mode) -> Mode

Return the port mode of self, given a streamlet interface mode.

Loading content...