Struct tydi::generator::common::Record

pub struct Record { /* fields omitted */ }

Inner struct for Type::Record

Methods

impl Record

pub fn new(name: impl Into<String>, fields: Vec<Field>) -> Record

Construct a new record.

pub fn new_empty(name: impl Into<String>) -> Record

Construct a new record without any fields.

pub fn new_empty_stream(name: impl Into<String>) -> Record

Construct a new record with a valid and ready bit.

pub fn insert_new_field(
&mut self,
name: impl Into<String>,
typ: Type,
reversed: bool
)

Create a new field and add it to the record.

pub fn insert(&mut self, field: Field)

Add a field to the record.

pub fn has_reversed_field(&self) -> bool

Returns true if the record contains a field that is reversed. Does not include nested records.

pub fn has_reversed(&self) -> bool

Returns true if the record contains a field that is reversed, including any nested records.

pub fn fields(&self) -> impl Iterator<Item = &Field>

Returns an iterable over the fields.

pub fn is_empty(&self) -> bool

Returns true if record contains no fields.

pub fn append_name_nested(&self, with: impl Into<String>) -> Self

Append a string to the name of this record, and any nested records.