Trait tydi::generator::vhdl::Split

pub trait Split {
    fn split(&self) -> (Option<Self>, Option<Self>)
    where
        Self: Sized;
}

Trait used to split types, ports, and record fields into a VHDL-friendly versions, since VHDL does not support bundles of wires with opposite directions.

Required methods

fn split(&self) -> (Option<Self>, Option<Self>) where
Self: Sized,

Split up self into a (downstream/forward, upstream/reverse) version, if applicable.