Module tydi::logical

Logical streams.

Reference

Structs

ElementStream

An element stream with a path name and LogicalType. Contains no nested streams.

Group

The Group stream type acts as a product type (composition).

Signals

Stream

The stream-manipulating logical stream type.

Union

Reference

Enums

Direction

Direction of a stream.

LogicalSplitItem

A split item is either an async signal (outside streamspace) or an element stream (no nested streams).

LogicalType

Types of logical streams.

PhysicalSplitItem

A split item is either an async signal (outside streamspace) or a physical stream.

Synchronicity

The synchronicity of the elements in the child stream with respect to the elements in the parent stream.