Enum tydi::logical::LogicalType

pub enum LogicalType {
    Null,
    Bits(Positive),
    Group(Group),
    Union(Union),
    Stream(Stream),
}

Types of logical streams.

This structure is at the heart of the logical stream specification. It is used both to specify the type of a logical stream and internally for the process of lowering the recursive structure down to physical streams and signals.

The logical stream type is defined recursively by means of a number of stream types. Two classes of stream types are defined: stream-manipulating types, and element-manipulating types.

Examples

Reference

Variants

Null

The Null stream type indicates the transferrence of one-valued data: it is only valid value is ∅ (null).

Reference

Bits(Positive)

Tuple Fields

0: Positive

The Bits stream type, defined as Bits(b), indicates the transferrence of 2^b-valued data carried by means of a group of b bits, whereb is a positive integer.

Reference

Group(Group)

Tuple Fields

0: Group

The Group stream type acts as a product type (composition).

Reference

Union(Union)

Tuple Fields

0: Union

The Union stream type acts as a sum type (exclusive disjunction).

Reference

Stream(Stream)

Tuple Fields

0: Stream

The Stream type is used to define a new physical stream.

Reference

Implementations

impl LogicalType

pub fn try_new_bits(bit_count: NonNegative) -> Result<Self>

Returns a new Bits stream type with the provided bit count as number of bits. Returns an error when the bit count is zero.

Examples

use tydi::{Error, logical::LogicalType, Positive};

let bits = LogicalType::try_new_bits(4);
let zero = LogicalType::try_new_bits(0);

assert_eq!(bits, Ok(LogicalType::Bits(Positive::new(4).unwrap())));
assert_eq!(zero, Err(Error::InvalidArgument("bit count cannot be zero".to_string())));

pub fn try_new_group(
group: impl IntoIterator<Item = (impl TryInto<Name, Error = impl Into<Box<dyn Error>>>, impl TryInto<LogicalType, Error = impl Into<Box<dyn Error>>>)>
) -> Result<Self>

Returns a new Group stream type from the provided iterator of names and stream types. Returns an error when the values cannot be converted into valid names, or valid logical stream types as required by Group.

Examples

use tydi::{Error, logical::{Group, LogicalType}};

let group = LogicalType::try_new_group(
    vec![
        ("a", 4), // TryFrom<NonNegative> for LogicalType::Bits.
        ("b", 12),
    ]
)?;

assert!(match group {
    LogicalType::Group(_) => true,
    _ => false,
});

assert_eq!(
    LogicalType::try_new_group(vec![("1badname", 4)]),
    Err(Error::InvalidArgument("name cannot start with a digit".to_string()))
);
assert_eq!(
    LogicalType::try_new_group(vec![("good_name", 0)]),
    Err(Error::InvalidArgument("bit count cannot be zero".to_string()))
);

pub fn try_new_union(
union: impl IntoIterator<Item = (impl TryInto<Name, Error = impl Into<Box<dyn Error>>>, impl TryInto<LogicalType, Error = impl Into<Box<dyn Error>>>)>
) -> Result<Self>

pub fn is_element_only(&self) -> bool

Returns true if this logical stream consists of only element- manipulating stream types. This recursively checks all inner stream types.

Examples

use tydi::logical::LogicalType;

assert!(LogicalType::Null.is_element_only());
assert!(LogicalType::try_new_bits(3)?.is_element_only());

pub fn is_null(&self) -> bool

Returns true if and only if this logical stream does not result in any signals.

Reference

pub fn compatible(&self, other: &LogicalType) -> bool

pub fn split(&self) -> IntoIter<LogicalSplitItem>

pub fn physical(&self) -> IntoIter<PhysicalSplitItem>