Struct tydi::logical::Group

pub struct Group(_);

The Group stream type acts as a product type (composition).

Reference

Implementations

impl Group

pub fn try_new(
group: impl IntoIterator<Item = (impl TryInto<Name, Error = impl Into<Box<dyn Error>>>, impl TryInto<LogicalType, Error = impl Into<Box<dyn Error>>>)>
) -> Result<Self>

Returns a new Group logical stream type. Returns an error when either the name or logical stream type conversion fails, or when there are duplicate names.

pub fn iter(&self) -> impl Iterator<Item = (&Name, &LogicalType)>

Returns an iterator over the fields of the Group.