Module tydi::parser::nom

Nom-based parsers for Streamlet Definition Files.

Functions

bits

bool

comment

comment_delimited
Delimited comments, not meant for doc strings, so if it succeeds, it produces an empty str.

comment_doc
Line comment meant for doc strings.

comment_doc_block

comment_line
Line or eof delimited comment, not meant for doc string, so if it succeeds, it produces an empty str.

complexity

direction

doc

group

interface

list_of_streamlets

logical_stream_type

mode

name

null

stream

streamlet

synchronicity

take_until_newline_or_eof

union