Struct tydi::physical::Complexity

pub struct Complexity { /* private fields */ }

Interface complexity level.

This logical stream parameter specifies the guarantees a source makes about how elements are transferred. Equivalently, it specifies the assumptions a sink can safely make.

Examples

use tydi::physical::Complexity;

let c3 = Complexity::new_major(3);
let c30 = Complexity::new(vec![3, 0])?;
let c31 = Complexity::new(vec![3, 1])?;
let c4 = Complexity::new_major(4);

assert_eq!(c3, c30);
assert!(c3 < c31);
assert!(c31 < c4);

assert_eq!(c31.to_string(), "3.1");

Reference

Implementations

impl Complexity

pub fn new(level: impl IntoIterator<Item = NonNegative>) -> Result<Self>

Constructs a new Complexity with provided level. Returns an error when the provided level iterator is empty.

Examples

use tydi::physical::Complexity;

let c = Complexity::new(vec![1, 2, 3, 4])?;
assert!(Complexity::new(vec![]).is_err());

pub fn new_major(level: NonNegative) -> Self

Constructs a new Complexity with provided level as major version.

Examples

use tydi::physical::Complexity;

let c = Complexity::new_major(4);

assert_eq!(c, Complexity::new(vec![4])?);

pub fn level(&self) -> &[NonNegative]

Returns the level of this Complexity.

Examples

use tydi::physical::Complexity;

let c = Complexity::new(vec![3, 14])?;
assert_eq!(c.level(), &[3, 14]);

pub fn major(&self) -> NonNegative

Returns the major version of this Complexity level.

Examples

use tydi::physical::Complexity;

let c = Complexity::new(vec![3, 14])?;
assert_eq!(c.major(), 3);