Module tydi::stdlib::common::architecture::assignment

Modules

array_assignment

assign

assignment_from

bitvec

declare

flatten

impls

Structs

AssignDeclaration

Describing the declaration of an assignment

Assignment

An object can be assigned from another object or directly

ObjectAssignment

An object can be assigned a value or another object

Enums

AssignmentKind

An object can be assigned a value or from another object

DirectAssignment

Directly assigning a value or an entire Record/Array

FieldSelection

A VHDL assignment constraint

RangeConstraint

A VHDL range constraint

StdLogicValue

Possible values which can be assigned to std_logic

ValueAssignment

Directly assigning a value or an entire Record, corresponds to the Types defined in tydi::generator::common::Type

Traits

Assign