Trait tydi::stdlib::common::architecture::assignment::flatten::FlatLength

pub trait FlatLength {
    fn flat_length(&self) -> Result<u32>;

    fn flat_length_for(
        &self, 
        nested_fields: &Vec<FieldSelection>
    ) -> Result<u32>;
}

Required methods

fn flat_length(&self) -> Result<u32>

Returns the total length of the object when flattened

fn flat_length_for(&self, nested_fields: &Vec<FieldSelection>) -> Result<u32>

Returns the total length of a field of the object