Module tydi::generator::common

Common hardware representation.

The goal of this module is to define some common constructs seen in structural hardware generation that back-ends may or may not use.

Modules

convert
Generator methods and implementations for Tydi types.

Structs

Component
A component.

Field
A field for a Record.

Package
A library of components and types.

Parameter
A parameter for components.

Port
A port.

Project
A project with libraries

Record
Inner struct for Type::Record

Enums

Mode
Modes for ports.

Type
Hardware types.

Type Definitions

TypeBundle
Bundle of names and types. Useful to represent flattened types.