Module tydi::generator::common::convert

Generator methods and implementations for Tydi types.

The generator module is enabled by the generator feature flag.

Constants

CANON_SUFFIX
Suffix provided to the canonical representation of streamlet components.

Traits

Componentify
Trait to create common representation components from things in the canonical way and user-friendly way.

ModeFor
Trait that helps to determine the common representation port mode given a streamlet interface mode.

Packify
Trait to create common representation of a package.

Portify
Trait to create common representation ports from things in the canonical way and user-friendly way.

Projectify
Trait to create common representation of a project.

Typify
Trait to create common representation types from things in the canonical way and user-friendly way.